sql: move benchmark tests to a new bench package
